+++ tests/test_plotting.py
@@ -1,4 +1,3 @@
-import pytest
 from _helpers import *
 
 import trisicell as tsc