+++ trisicell/datasets/_datasets.py
@@ -62,11 +62,14 @@ def melanoma20():
 
 
 def colorectal1():
-    """Human Colorectal Cancer dataset patient 1.
+    """Human Colorectal Cancer (Patient 1).
 
     This dataset was introduced in :cite:`Leung_2017` and was used in:
 
     * :cite:`B-SCITE` Figure 8a.
+    * :cite:`SiFit` Figure 6.
+    * :cite:`SPhyR` Table 1.
+    * :cite:`SiCloneFit` Figure 3.
 
     The size is n_cells × n_muts = 72 × 12
 
@@ -76,17 +79,19 @@ def colorectal1():
         An anndata in which `.X` is the input noisy.
     """
 
-    # TODO: extract.
+    # TODO: extract (SiFit 178 × 16).
     pass
 
 
 def colorectal2():
-    """Human Colorectal Cancer dataset patient 2.
+    """Human Colorectal Cancer (Patient 2).
 
     This dataset was introduced in :cite:`Leung_2017` and was used in:
 
     * :cite:`PhISCS` Figure 7.
     * :cite:`B-SCITE` Figure 8b.
+    * :cite:`SiCloneFit` Figure 4.
+    * :cite:`SCARLET` Figure 4.
 
     The size is n_cells × n_muts = 78 × 25
 
@@ -103,18 +108,18 @@ def colorectal2():
     """
 
     adata = tsc.io.read(tsc.ul.get_file("trisicell.datasets/real/colorectal2.h5ad"))
-    # TODO: 86 x 25 in B-SCITE?
+    # TODO: (86 x 25 in B-SCITE) (182 x 36 SiCloneFit)
     return adata
 
 
-def acute_lymphocytic_leukemia1():
-    """Human Acute Lymphocytic Leukemia dataset.
+def colorectal3():
+    """Human Colorectal Cancer.
 
-    This dataset was introduced in :cite:`Gawad_2014` and was used in:
+    This dataset was introduced in :cite:`Wu_2016` and was used in:
 
-    * :cite:`B-SCITE` Figure 5.
+    * :cite:`SiFit` Figure 5.
 
-    The size is n_cells × n_muts = 111 × 20
+    The size is n_cells × n_muts = 48 × 77
 
     Returns
     -------
@@ -126,13 +131,38 @@ def acute_lymphocytic_leukemia1():
     pass
 
 
+def acute_lymphocytic_leukemia1():
+    """Human Acute Lymphocytic Leukemia dataset (Patient 1).
+
+    This dataset was introduced in :cite:`Gawad_2014` and was used in:
+
+    * :cite:`B-SCITE` Figure 5.
+    * :cite:`infSCITE` Figure S16.
+
+    The size is n_cells × n_muts = 111 × 20
+
+    Returns
+    -------
+    :class:`anndata.AnnData`
+        An anndata in which `.X` is the input noisy.
+    """
+
+    adata = tsc.io.read(
+        tsc.ul.get_file("trisicell.datasets/real/acute_lymphocytic_leukemia1.h5ad")
+    )
+    # TODO: extract.
+    return adata
+
+
 def acute_lymphocytic_leukemia2():
-    """Human Acute Lymphocytic Leukemia dataset.
+    """Human Acute Lymphocytic Leukemia dataset (Patient 2).
 
     This dataset was introduced in :cite:`Gawad_2014` and was used in:
 
     * :cite:`PhISCS` Figure 9.
     * :cite:`B-SCITE` in Figure 6.
+    * :cite:`infSCITE` Figure S17.
+    * :cite:`Phyolin` Table 2.
 
     The size is n_cells × n_muts = 102 × 16
 
@@ -154,23 +184,257 @@ def acute_lymphocytic_leukemia2():
 
 
 def acute_lymphocytic_leukemia3():
-    # TODO: extract.
-    pass
+    """Human Acute Lymphocytic Leukemia dataset (Patient 3).
+
+    This dataset was introduced in :cite:`Gawad_2014` and was used in:
+
+    * :cite:`infSCITE` Figure S18.
+    * :cite:`SCIPhI` Figure 5.
+    * :cite:`ScisTree` Figure S3.
+
+    The size is n_cells × n_muts = 150 × 49
+
+    Returns
+    -------
+    :class:`anndata.AnnData`
+        An anndata in which `.X` is the input noisy.
+    """
+
+    adata = tsc.io.read(
+        tsc.ul.get_file("trisicell.datasets/real/acute_lymphocytic_leukemia3.h5ad")
+    )
+    # TODO: 255 x 49 in SCIPhI and ScisTree?
+    return adata
 
 
 def acute_lymphocytic_leukemia4():
+    """Human Acute Lymphocytic Leukemia dataset (Patient 4).
+
+    This dataset was introduced in :cite:`Gawad_2014` and was used in:
+
+    * :cite:`infSCITE` Figure S19.
+    * :cite:`gpps` Figure 3.
+    * :cite:`SASC` Figure 6.
+
+    The size is n_cells × n_muts = 143 × 78
+
+    Returns
+    -------
+    :class:`anndata.AnnData`
+        An anndata in which `.X` is the input noisy.
+    """
+
+    adata = tsc.io.read(
+        tsc.ul.get_file("trisicell.datasets/real/acute_lymphocytic_leukemia4.h5ad")
+    )
     # TODO: extract.
-    pass
+    return adata
 
 
 def acute_lymphocytic_leukemia5():
+    """Human Acute Lymphocytic Leukemia dataset (Patient 5).
+
+    This dataset was introduced in :cite:`Gawad_2014` and was used in:
+
+    * :cite:`infSCITE` Figure S20.
+    * :cite:`SASC` Figure 7.
+
+    The size is n_cells × n_muts = 96 × 105
+
+    Returns
+    -------
+    :class:`anndata.AnnData`
+        An anndata in which `.X` is the input noisy.
+    """
+
+    adata = tsc.io.read(
+        tsc.ul.get_file("trisicell.datasets/real/acute_lymphocytic_leukemia5.h5ad")
+    )
     # TODO: extract.
-    pass
+    return adata
 
 
 def acute_lymphocytic_leukemia6():
+    """Human Acute Lymphocytic Leukemia dataset (Patient 6).
+
+    This dataset was introduced in :cite:`Gawad_2014` and was used in:
+
+    * :cite:`infSCITE` Figure S21.
+    * :cite:`Phyolin` Table 2 and Figure 5.
+
+    The size is n_cells × n_muts = 146 × 10
+
+    Returns
+    -------
+    :class:`anndata.AnnData`
+        An anndata in which `.X` is the input noisy.
+    """
+
+    adata = tsc.io.read(
+        tsc.ul.get_file("trisicell.datasets/real/acute_lymphocytic_leukemia6.h5ad")
+    )
     # TODO: extract.
-    pass
+    return adata
+
+
+def tnbc():
+    """Triple-negative Breast Cancer.
+
+    This dataset was introduced in :cite:`Wang_2014` and was used in:
+
+    * :cite:`SCIPhI` Figure 4.
+    * :cite:`B-SCITE` Figure 7.
+    * :cite:`TRaIT` Figure 6.
+
+    The size is n_cells × n_muts = 16 × 20
+
+    Returns
+    -------
+    :class:`anndata.AnnData`
+        An anndata in which `.X` is the input noisy.
+
+    Examples
+    --------
+    >>> adata = tsc.datasets.tnbc()
+    >>> df_in = adata.to_df()
+    """
+
+    adata = tsc.io.read(tsc.ul.get_file("trisicell.datasets/real/tnbc.h5ad"))
+    return adata
+
+
+def erbc():
+    """Oestrogen-receptor-positive (ER+) Breast Cancer.
+
+    This dataset was introduced in :cite:`Wang_2014` and was used in:
+
+    * :cite:`SCITE` Figure S8 and S9.
+    * :cite:`infSCITE` Figure S15.
+    * :cite:`gpps` Figure 1.
+
+    The size is n_cells × n_muts = 47 × 40
+
+    Returns
+    -------
+    :class:`anndata.AnnData`
+        An anndata in which `.X` is the input noisy.
+
+            - `.uns['params_scite']` is parameters inferred by SCITE.
+
+    Examples
+    --------
+    >>> adata = tsc.datasets.erbc()
+    >>> df_in = adata.to_df()
+    """
+
+    adata = tsc.io.read(tsc.ul.get_file("trisicell.datasets/real/erbc.h5ad"))
+    return adata
+
+
+def muscle_invasive_bladder():
+    """Muscle Invasive Bladder Cancer.
+
+    This dataset was introduced in :cite:`Li_2012` and was used in:
+
+    * :cite:`OncoNEM` Figure 6B.
+
+    The size is n_cells × n_muts = 44 × 443
+
+    Returns
+    -------
+    :class:`anndata.AnnData`
+        An anndata in which `.X` is the input noisy.
+
+            - `.uns['params_onconem']` is parameters inferred by OncoNEM.
+
+    Examples
+    --------
+    >>> adata = tsc.datasets.muscle_invasive_bladder()
+    >>> df_in = adata.to_df()
+    """
+
+    adata = tsc.io.read(
+        tsc.ul.get_file("trisicell.datasets/real/muscle_invasive_bladder.h5ad")
+    )
+    return adata
+
+
+def renal_cell_carcinoma():
+    """Clear-cell Renal-cell Carcinoma.
+
+    This dataset was introduced in :cite:`Xu_2012` and was used in:
+
+    * :cite:`SCITE` Figure S6 and S7.
+    * :cite:`infSCITE` Figure S14.
+
+    The size is n_cells × n_muts = 17 × 35
+
+    Returns
+    -------
+    :class:`anndata.AnnData`
+        An anndata in which `.X` is the input noisy.
+
+            - `.uns['params_scite']` is parameters inferred by SCITE.
+
+    Examples
+    --------
+    >>> adata = tsc.datasets.renal_cell_carcinoma()
+    >>> df_in = adata.to_df()
+    """
+
+    adata = tsc.io.read(
+        tsc.ul.get_file("trisicell.datasets/real/renal_cell_carcinoma.h5ad")
+    )
+    return adata
+
+
+def myeloproliferative_neoplasms712():
+    """JAK2-Negative Myeloproliferative Neoplasm.
+
+    This dataset was introduced in :cite:`Hou_2012` and was used in:
+
+    * :cite:`OncoNEM` Figure 6D.
+
+    The size is n_cells × n_muts = 58 × 712
+
+    Returns
+    -------
+    :class:`anndata.AnnData`
+        An anndata in which `.X` is the input noisy.
+
+            - `.uns['params_onconem']` is parameters inferred by OncoNEM.
+    """
+
+    adata = tsc.io.read(
+        tsc.ul.get_file("trisicell.datasets/real/myeloproliferative_neoplasms712.h5ad")
+    )
+    return adata
+
+
+def myeloproliferative_neoplasms78():
+    """JAK2-Negative Myeloproliferative Neoplasm.
+
+    This dataset was introduced in :cite:`Hou_2012` and was used in:
+
+    * :cite:`SCITE` Figure S5.
+
+    The size is n_cells × n_muts = 58 × 78
+
+    Returns
+    -------
+    :class:`anndata.AnnData`
+        An anndata in which `.X` is the input noisy.
+
+    Notes
+    -----
+    The original dataset contains 712 mutations but 78 ones were considered as
+    non-synonymous mutations from the full data.
+    """
+
+    adata = tsc.io.read(
+        tsc.ul.get_file("trisicell.datasets/real/myeloproliferative_neoplasms78.h5ad")
+    )
+    return adata
 
 
 def myeloproliferative_neoplasms18():
@@ -178,7 +442,10 @@ def myeloproliferative_neoplasms18():
 
     This dataset was introduced in :cite:`Hou_2012` and was used in:
 
-    * :cite:`PhISCS` Figure 9.
+    * :cite:`SCITE` Figure S2, S3 and S4.
+    * :cite:`Kim_2014` Figure 1.
+    * :cite:`infSCITE` Figure S13.
+    * :cite:`gpps` Figure 2.
 
     The size is n_cells × n_muts = 58 × 18
 
@@ -187,21 +454,28 @@ def myeloproliferative_neoplasms18():
     :class:`anndata.AnnData`
         An anndata in which `.X` is the input noisy.
 
-            - `.layers['solution_fig9']` is the solution presented in Figure 9 of PhISCS paper.
-            - `.uns['params_fig9']` is parameters used as input to get 'solution_fig9'.
-            - `.var` includes information of the bulk samples.
+            - `.uns['params_scite']` is parameters inferred by SCITE.
+
+    Notes
+    -----
+    The original dataset contains 712 mutations but 18 ones were considered as
+    cancer related mutations from the full data.
     """
 
-    pass
+    adata = tsc.io.read(
+        tsc.ul.get_file("trisicell.datasets/real/myeloproliferative_neoplasms18.h5ad")
+    )
+    return adata
 
 
-def triple_negative_breast_cancer():
-    """Triple-negative Breast Cancer.
+def high_grade_serous_ovarian_cancer1():
+    """Triple-negative Breast Cancer (Patient 2).
 
-    This dataset was introduced in []_ and was used in:
-    * :cite:`B-SCITE` Figure 7.
+    This dataset was introduced in :cite:`McPherson_2016` and was used in:
 
-    The size is n_cells × n_muts = 16 × 18
+    * :cite:`infSCITE` Figure S22.
+
+    The size is n_cells × n_muts = 588 × 37
 
     Returns
     -------
@@ -209,12 +483,65 @@ def triple_negative_breast_cancer():
         An anndata in which `.X` is the input noisy.
     """
 
+    adata = tsc.io.read(
+        tsc.ul.get_file(
+            "trisicell.datasets/real/high_grade_serous_ovarian_cancer1.h5ad"
+        )
+    )
     # TODO: extract.
-    pass
+    return adata
 
 
-def high_grade_serous_ovarian_cancer():
-    """Triple-negative Breast Cancer.
+def high_grade_serous_ovarian_cancer2():
+    """Triple-negative Breast Cancer (Patient 3).
+
+    This dataset was introduced in :cite:`McPherson_2016` and was used in:
+
+    * :cite:`infSCITE` Figure S23.
+
+    The size is n_cells × n_muts = 672 × 60
+
+    Returns
+    -------
+    :class:`anndata.AnnData`
+        An anndata in which `.X` is the input noisy.
+    """
+
+    adata = tsc.io.read(
+        tsc.ul.get_file(
+            "trisicell.datasets/real/high_grade_serous_ovarian_cancer2.h5ad"
+        )
+    )
+    # TODO: extract.
+    return adata
+
+
+def high_grade_serous_ovarian_cancer3():
+    """Triple-negative Breast Cancer (Patient 9).
+
+    This dataset was introduced in :cite:`McPherson_2016` and was used in:
+
+    * :cite:`infSCITE` Figure S24.
+
+    The size is n_cells × n_muts = 420 × 37
+
+    Returns
+    -------
+    :class:`anndata.AnnData`
+        An anndata in which `.X` is the input noisy.
+    """
+
+    adata = tsc.io.read(
+        tsc.ul.get_file(
+            "trisicell.datasets/real/high_grade_serous_ovarian_cancer3.h5ad"
+        )
+    )
+    # TODO: extract.
+    return adata
+
+
+def high_grade_serous_ovarian_cancer_3celllines():
+    """Triple-negative Breast Cancer ().
 
     This dataset was introduced in :cite:`Laks_2019`
     and preprocessed in :cite:`McPherson_2019`.
@@ -243,7 +570,7 @@ def high_grade_serous_ovarian_cancer():
 
     Examples
     --------
-    >>> sc = tsc.datasets.high_grade_serous_ovarian_cancer()
+    >>> sc = tsc.datasets.high_grade_serous_ovarian_cancer_3celllines()
     >>> print(sc)
     AnnData object with n_obs × n_vars = 891 × 13666
         obs: 'clone_id', 'group_color', 'cell_name'