+++ trisicell/tl/__init__.py
@@ -1,5 +1,6 @@
 from trisicell.tl.cna import infercna
 from trisicell.tl.consensus import consensus_combine, consensus_tree
 from trisicell.tl.partition_function import partition_function
+from trisicell.tl.sbm import sbm
 from trisicell.tl.score import *
 from trisicell.tl.solver import *